fix recording window jiggle, add debug mode to mac build scripts

+++ mac/make.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
 cd /Volumes/Two/anki
-echo "cleaning up..."
-rm -rf build dist
+if [ "x$debug" = "x" ]; then
+    echo "cleaning up..."
+    rm -rf build dist
+fi
 echo "syncing updates..."
 rsync -av reflex:Lib/code/libanki --exclude .git --exclude build --exclude dist --delete .
 rsync -av reflex:Lib/code/ankiqt --exclude .git --exclude build --exclude dist  --delete .
